chore(rules): move loader back to setup/match.go
