--- tool/internal/rule/loader_test.go
@@ -1,144 +1,0 @@
-// Copyright The OpenTelemetry Authors
-// SPDX-License-Identifier: Apache-2.0
-
-package rule
-
-import (
-	"testing"
-
-	"gopkg.in/yaml.v3"
-)
-
-func TestCreateRuleFromFields(t *testing.T) {
-	tests := []struct {
-		name         string
-		yamlContent  string
-		ruleName     string
-		expectError  bool
-		expectedType string
-	}{
-		{
-			name: "struct rule creation",
-			yamlContent: `
-struct: TestStruct
-target: github.com/example/lib
-`,
-			ruleName:     "test-struct-rule",
-			expectError:  false,
-			expectedType: "*InstStructRule",
-		},
-		{
-			name: "func rule creation",
-			yamlContent: `
-func: TestFunc
-target: github.com/example/lib
-before: MyHook1Before
-`,
-			ruleName:     "test-func-rule",
-			expectError:  false,
-			expectedType: "*InstFuncRule",
-		},
-		{
-			name: "file rule creation",
-			yamlContent: `
-file: test.go
-target: github.com/example/lib
-`,
-			ruleName:     "test-file-rule",
-			expectError:  false,
-			expectedType: "*InstFileRule",
-		},
-		{
-			name: "raw rule creation",
-			yamlContent: `
-raw: test
-target: github.com/example/lib
-`,
-			ruleName:     "test-raw-rule",
-			expectError:  false,
-			expectedType: "*InstRawRule",
-		},
-		{
-			name: "rule with version",
-			yamlContent: `
-struct: TestStruct
-target: github.com/example/lib
-version: v1.0.0,v2.0.0
-`,
-			ruleName:     "test-versioned-rule",
-			expectError:  false,
-			expectedType: "*InstStructRule",
-		},
-		{
-			name: "invalid yaml syntax",
-			yamlContent: `
-struct: [
-target: github.com/example/lib
-`,
-			ruleName:    "test-invalid-rule",
-			expectError: true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			testCreateRuleFromFieldsCase(t, tt)
-		})
-	}
-}
-
-func testCreateRuleFromFieldsCase(t *testing.T, tt struct {
-	name         string
-	yamlContent  string
-	ruleName     string
-	expectError  bool
-	expectedType string
-},
-) {
-	var fields map[string]any
-	err := yaml.Unmarshal([]byte(tt.yamlContent), &fields)
-	if err != nil {
-		if !tt.expectError {
-			t.Fatalf("failed to parse test YAML: %v", err)
-		}
-		return // Expected YAML parsing to fail
-	}
-
-	createdRule, err := CreateRuleFromFields([]byte(tt.yamlContent), tt.ruleName, fields)
-
-	if tt.expectError {
-		if err == nil {
-			t.Error("expected error but got none")
-		}
-		return
-	}
-
-	if err != nil {
-		t.Errorf("unexpected error: %v", err)
-		return
-	}
-
-	if createdRule == nil {
-		return
-	}
-
-	validateCreatedRule(t, createdRule, tt.ruleName, fields)
-}
-
-func validateCreatedRule(t *testing.T, createdRule InstRule, ruleName string, fields map[string]any) {
-	if createdRule.GetName() != ruleName {
-		t.Errorf("rule name = %v, want %v", createdRule.GetName(), ruleName)
-	}
-
-	if target, ok := fields["target"].(string); ok {
-		if createdRule.GetTarget() != target {
-			t.Errorf("rule target = %v, want %v", createdRule.GetTarget(), target)
-		}
-	}
-
-	if version, ok := fields["version"].(string); ok {
-		if createdRule.GetVersion() != version {
-			t.Errorf("rule version = %v, want %v", createdRule.GetVersion(), version)
-		}
-	}
-}
